_code-samples/claim-payment-channel/py/requirements.txt
xrpl-py>=2.0.0
